Don't require a remote kubeconfig
Instead, require that at least one kubeconfig file is provided. This allows AK
to be deployed as a pod in either the local or the remote cluster, but ensures
that both ends won't fall back to in-cluster config (i.e. that AK won't use the
same cluster as local and remote).

Signed-off-by: Nic Cope <negz@rk0n.org>